Create Capture & Translate GCD value in console
This scripts serves as a way to translate new GCD parameter (more granular than old gcl) in order to make it readable and understandable for digital marketers, to soften the debugging efforts in terms of consent. Although it is still in progress and there is not clear information about the gcd values from google, I´ve investigated and used various information across the internet to somehow make it reliable. I will try to update it and improve it with time, but so far it seems to work decently